+++ packages/runtime/src/widgets/Charts/StackBarChart.tsx
@@ -2,9 +2,10 @@ import { Bar } from "react-chartjs-2";
 import type { ChartOptions } from "chart.js";
 import { useRegisterBindings } from "@ensembleui/react-framework";
 import { useState } from "react";
-import type { ChartDataSets, ChartProps } from "..";
 import ChartDataLabels from "chartjs-plugin-datalabels";
 import { get } from "lodash-es";
+import { type ChartDataSets, type ChartProps } from "..";
+import { getMergedOptions } from "./utils/getMergedOptions";
 
 const options: ChartOptions<"bar"> = {
   maintainAspectRatio: false,
@@ -65,16 +66,7 @@ export const StackBarChart: React.FC<ChartProps> = (props) => {
         labels: values?.labels,
         datasets: config?.data?.datasets as ChartDataSets[],
       }}
-      options={{
-        ...options,
-        plugins: {
-          ...options.plugins,
-          title: {
-            display: Boolean(values?.title),
-            text: values?.title,
-          },
-        },
-      }}
+      options={getMergedOptions(options, values?.title, config?.options)}
       plugins={[ChartDataLabels]}
       style={{
         ...(get(props, "styles") as object),