+++ packages/runtime/src/widgets/index.ts
@@ -27,3 +27,5 @@ export * from "./Stepper";
 export * from "./ToggleButton";
 export * from "./Date";
 export * from "./SignInWithGoogle";
+export * from "./LoadingContainer";
+export * from "./Skeleton";